refactor(api): extract equalWithGetter from test-helpers.js
