+++ api/tests/tooling/chai-custom-helpers/index.js
@@ -1,7 +1,15 @@
 import { deepEqualArray } from './deep-equal-array.js';
 import { deepEqualInstance } from './deep-equal-instance.js';
 import { deepEqualInstanceOmitting } from './deep-equal-instance-omitting.js';
+import { equalWithGetter } from './equal-with-getter.js';
 import { exactlyContain } from './exactly-contain.js';
 import { exactlyContainInOrder } from './exactly-contain-in-order.js';
 
-export { deepEqualArray, deepEqualInstance, deepEqualInstanceOmitting, exactlyContain, exactlyContainInOrder };
+export {
+  deepEqualArray,
+  deepEqualInstance,
+  deepEqualInstanceOmitting,
+  equalWithGetter,
+  exactlyContain,
+  exactlyContainInOrder,
+};
